requirements.txt
# Python
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Machine Learning
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0

# Database
psycopg2-binary==2.9.9

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-decouple==3.8
